telegram-be/pyproject.toml
[project]
name = "telegram-dashboard"
version = "0.1.0"
description = "Telegram Chat Integration Dashboard - Phase 1"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "telethon>=1.36.0",
    "supabase>=2.9.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "websockets>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []